patch 8.1.2319: compiler warning for int size
Problem:    Compiler warning for int size.
Solution:   Add typecast. (Mike Williams)

+++ src/mouse.c
@@ -2884,7 +2884,7 @@ mouse_comp_pos(
 	col += win->w_skipcol;
 	// limit to text length plus one
 	p = ml_get_buf(win->w_buffer, lnum, FALSE);
-	count = STRLEN(p);
+	count = (int)STRLEN(p);
 	if (col > count)
 	    col = count;
     }
